feat(graph): Add versioning and migration system for graphs
Implements graph versioning with automatic migration to latest schema:

- Move graph utility functions to separate file
- Add version field ("2024-12-09") to graph schema
- Create migration system to handle legacy graphs
- Add version check and auto-migration on graph load
- Relocate and update related tests

BREAKING CHANGE: Graphs now require version field in schema
